fix: update imports after kebab-case renames and remove Storybook
Fix broken module paths in tests and app code, drop Storybook config and stories, and remove STORYBOOK-only mock paths.
