--- clients/web/.storybook/main.ts
@@ -1,22 +1,0 @@
-import type { StorybookConfig } from '@storybook/react-vite'
-import tailwindcss from '@tailwindcss/vite'
-
-const config: StorybookConfig = {
-  stories: ['../src/**/*.stories.@(js|jsx|mjs|ts|tsx)'],
-  addons: ['@storybook/addon-a11y', '@storybook/addon-docs'],
-  framework: {
-    name: '@storybook/react-vite',
-    options: {},
-  },
-  async viteFinal(cfg) {
-    const { mergeConfig } = await import('vite')
-    return mergeConfig(cfg, {
-      plugins: [tailwindcss()],
-      define: {
-        'import.meta.env.STORYBOOK': JSON.stringify('true'),
-      },
-    })
-  },
-}
-
-export default config
