--- clients/web/.storybook/preview.ts
@@ -1,16 +1,0 @@
-import type { Preview } from '@storybook/react-vite'
-import '../src/index.css'
-
-const preview: Preview = {
-  parameters: {
-    layout: 'padded',
-    controls: {
-      matchers: {
-        color: /(background|color)$/i,
-        date: /Date$/i,
-      },
-    },
-  },
-}
-
-export default preview
