+++ clients/web/package.json
@@ -15,8 +15,6 @@
     "test:tdd": "vitest",
     "test:coverage": "vitest run --coverage",
     "openapi:types": "openapi-typescript http://127.0.0.1:8080/api/openapi.json -o src/lib/generated/openapi-types.ts",
-    "storybook": "STORYBOOK_DISABLE_TELEMETRY=1 storybook dev -p 6006",
-    "build-storybook": "STORYBOOK_DISABLE_TELEMETRY=1 storybook build -o storybook-static",
     "analyze": "ANALYZE=1 vite build",
     "bundle:check": "node scripts/check-bundle-size.mjs"
   },
@@ -63,9 +61,6 @@
   },
   "devDependencies": {
     "@eslint/js": "^9.39.4",
-    "@storybook/addon-a11y": "^10.3.5",
-    "@storybook/addon-docs": "^10.3.5",
-    "@storybook/react-vite": "^10.3.5",
     "@tailwindcss/vite": "^4.2.2",
     "@testing-library/dom": "^10.4.1",
     "@testing-library/jest-dom": "^6.9.1",
@@ -84,7 +79,6 @@
     "eslint-plugin-jsx-a11y": "^6.10.2",
     "eslint-plugin-react-hooks": "^7.0.1",
     "eslint-plugin-react-refresh": "^0.5.2",
-    "eslint-plugin-storybook": "10.3.5",
     "fake-indexeddb": "^6.2.5",
     "globals": "^17.4.0",
     "husky": "^9.1.7",
@@ -95,7 +89,6 @@
     "playwright": "^1.59.1",
     "remark-parse": "^11.0.0",
     "rollup-plugin-visualizer": "^5.14.0",
-    "storybook": "^10.3.5",
     "tailwindcss": "^4.2.2",
     "typescript": "~6.0.2",
     "typescript-eslint": "^8.58.0",
