+++ clients/web/src/app.tsx
@@ -54,7 +54,7 @@ import PrivacyPolicyHistoryPage from './pages/privacy-policy-history-page'
 import MagicLinkPage from './pages/magic-link'
 import ResetPassword from './pages/reset-password'
 import Signup from './pages/signup'
-import ParentDashboard from './pages/lms/parent/ParentDashboard'
+import ParentDashboard from './pages/lms/parent/parent-dashboard'
 import TermsOfUsePage from './pages/terms-of-use-page'
 import TermsOfUseHistoryPage from './pages/terms-of-use-history-page'
 import TrustCenterPage from './pages/trust-center-page'
