+++ clients/web/src/components/editor/math-insert-popover.tsx
@@ -2,5 +2,5 @@
  * @deprecated Use EquationEditorProvider + EquationEditorDialog (feature 8.11).
  * Re-exported for backward compatibility.
  */
-export { EquationEditorDialog as MathInsertPopover } from './EquationEditorDialog'
-export type { EquationEditorDialogProps as MathInsertPopoverProps } from './EquationEditorDialog'
+export { EquationEditorDialog as MathInsertPopover } from './equation-editor-dialog'
+export type { EquationEditorDialogProps as MathInsertPopoverProps } from './equation-editor-dialog'
