+++ clients/web/src/components/settings/notification-preferences-panel.tsx
@@ -29,12 +29,6 @@ const DIGEST_OPTIONS: { value: PreferenceRow['digestMode']; label: string }[] = 
   { value: 'off', label: 'Off' },
 ]
 
-const STORYBOOK_MOCK_ROWS: PreferenceRow[] = [
-  { eventType: 'grade_posted', emailEnabled: true, pushEnabled: true, digestMode: 'instant' },
-  { eventType: 'assignment_created', emailEnabled: true, pushEnabled: true, digestMode: 'daily' },
-  { eventType: 'discussion_reply', emailEnabled: false, pushEnabled: true, digestMode: 'off' },
-]
-
 export function NotificationPreferencesPanel() {
   const baseId = useId()
   const [rows, setRows] = useState<PreferenceRow[]>([])
@@ -44,11 +38,6 @@ export function NotificationPreferencesPanel() {
   const [pushLoading, setPushLoading] = useState(false)
 
   const load = useCallback(async () => {
-    if (import.meta.env.STORYBOOK === 'true') {
-      setRows(STORYBOOK_MOCK_ROWS)
-      setLoading(false)
-      return
-    }
     setLoading(true)
     try {
       const res = await authorizedFetch('/api/v1/me/notification-preferences')
@@ -67,9 +56,7 @@ export function NotificationPreferencesPanel() {
 
   useEffect(() => {
     void load()
-    if (import.meta.env.STORYBOOK !== 'true') {
-      void getExistingPushSubscription().then((sub) => setPushSubscribed(!!sub))
-    }
+    void getExistingPushSubscription().then((sub) => setPushSubscribed(!!sub))
   }, [load])
 
   const updateRow = (eventType: string, patch: Partial<PreferenceRow>) => {
