+++ clients/web/src/pages/lms/module-external-link-modal.tsx
@@ -5,7 +5,7 @@ import {
   OneDrivePicker,
   DropboxPicker,
   type PickedFile,
-} from '../../services/cloudPicker'
+} from '../../services/cloud-picker'
 
 type ModuleExternalLinkModalProps = {
   open: boolean
