+++ docs/plan/10-compliance-privacy-security/10.7-wcag-conformance-program.md
@@ -63,7 +63,7 @@ Section 508 of the Rehabilitation Act (29 U.S.C. § 794d) requires all US federa
 - **Accessibility** — This plan IS the accessibility conformance program; no separate note needed.
 - **Scalability** — Accessibility CI gate scales with existing GitHub Actions/CI infrastructure.
 - **Reliability** — axe-core checks run on every PR; failures block merge.
-- **Observability** — Axe violation counts tracked per component; Storybook accessibility add-on for design-review stage.
+- **Observability** — Axe violation counts tracked per component; ESLint `jsx-a11y` and Vitest/Playwright scans for design-review stage.
 - **Maintainability** — Shared `clients/web/src/lib/a11y/` utility for focus management, ARIA announcer, skip-link component.
 - **Internationalization** — All ARIA labels and error messages MUST use i18n keys (11.1); RTL layout must maintain logical focus order (11.2).
 - **Backward compatibility** — Accessibility fixes are additive (ARIA attributes, semantic HTML); no functional regressions.
@@ -93,7 +93,7 @@ No new API endpoints. The conformance statement is a static page. Axe-core resul
 - **Color tokens** — all design tokens in `clients/web/src/styles/` validated against WCAG 4.5:1 normal text, 3:1 large text, 3:1 UI component contrast.
 - **Keyboard shortcut reference** — accessible modal documenting shortcuts (triggered by `?`).
 - **Conformance statement page** (`/accessibility`) — public, updated annually.
-- All new components in Storybook MUST have accessibility story with axe check.
+- New components MUST have Vitest coverage and pass axe checks in CI where applicable.
 
 ## 11. AI / ML Considerations
 
@@ -105,7 +105,7 @@ None. However, the AI-generated alt text feature (12.5) depends on the infrastru
 - `clients/web/src/components/` — all existing components audited.
 - `clients/web/src/lib/a11y/` — new shared utilities.
 - CI configuration (GitHub Actions or equivalent) — axe-core Playwright integration.
-- Storybook — `@storybook/addon-a11y` installed.
+- `clients/web` — ESLint `jsx-a11y` rules; axe-core via Playwright in CI.
 - Cross-link: Plan 12.1–12.10 (Section 12 Accessibility) — all Section 12 plans depend on this conformance program as the overarching standard.
 - Cross-link: Plan 10.8 (VPAT) — this plan's audit output feeds the VPAT.
 
@@ -120,7 +120,7 @@ None. However, the AI-generated alt text feature (12.5) depends on the infrastru
 |---|---|---|---|
 | TipTap editor has deep accessibility gaps requiring upstream contributions | H | H | Evaluate Tiptap v2 accessibility roadmap; consider Prosemirror wrapper patches; budget 4w for editor alone |
 | axe-core misses manual-only WCAG criteria | H | M | Supplement with manual screen-reader testing (NVDA/Chrome, JAWS/Edge, VoiceOver/Safari) quarterly |
-| Design token changes break contrast compliance | M | M | Automated contrast-ratio linting in Storybook and CI |
+| Design token changes break contrast compliance | M | M | Automated contrast-ratio linting in CI |
 | Accessibility debt accumulates faster than remediation | M | H | Strict axe gate on new PRs; dedicated 20% sprint capacity for a11y remediation |
 
 ## 15. Rollout Plan
@@ -145,7 +145,7 @@ None. However, the AI-generated alt text feature (12.5) depends on the infrastru
 ## 17. Documentation & Training
 
 - Developer guide: "Accessibility standards and component-level annotations."
-- Component Storybook: accessibility story template for all new components.
+- Component tests: accessibility assertions in Vitest/Playwright for new components.
 - Public: Accessibility Conformance Statement page at `/accessibility`.
 - Training: 2-hour accessibility workshop for frontend engineers and designers.
