+++ docs/plan/11-i18n-l10n/11.2-rtl-support.md
@@ -51,7 +51,7 @@ Arabic, Hebrew, Persian (Farsi), Urdu, and Pashto are right-to-left (RTL) langua
 - **FR-4.** The TipTap editor MUST support RTL text direction: the editor's content node MUST have `dir="auto"` (Unicode bidi algorithm); the toolbar MUST mirror directional buttons (indent/outdent, list alignment) when RTL is active.
 - **FR-5.** All SVG icons with directional meaning (arrow-right, chevron-right, back button) MUST be mirrored using `transform: scaleX(-1)` or the Tailwind `rtl:rotate-180` variant when in RTL mode.
 - **FR-6.** The gradebook grid, module sidebar, and navigation breadcrumbs MUST render correctly in RTL layout.
-- **FR-7.** The system MUST NOT break LTR layout when RTL support is added; Chromatic or Storybook visual regression tests MUST confirm no LTR regressions.
+- **FR-7.** The system MUST NOT break LTR layout when RTL support is added; Playwright visual regression tests MUST confirm no LTR regressions.
 
 ## 6. Non-Functional Requirements
 
@@ -62,7 +62,7 @@ Arabic, Hebrew, Persian (Farsi), Urdu, and Pashto are right-to-left (RTL) langua
 - **Scalability** — CSS logical properties are a compile-time change; no runtime scaling concern.
 - **Reliability** — Locale switch must not require page reload; RTL/LTR toggle must be instant.
 - **Observability** — No special metrics needed; RTL locale selection tracked by standard `locale_active{locale}` metric from i18n framework (11.1).
-- **Maintainability** — ESLint rule (`eslint-plugin-i18next` or custom) flags physical CSS property usage in new components. Storybook RTL story required for all layout components.
+- **Maintainability** — ESLint rule (`eslint-plugin-i18next` or custom) flags physical CSS property usage in new components. Vitest/Playwright RTL coverage required for layout components.
 - **Internationalization** — This plan IS the RTL i18n concern. Tailwind v4 `rtl:` variant documentation linked in developer guide.
 - **Backward compatibility** — LTR locales unchanged; no CSS changes visible to LTR users if logical properties are correctly equivalent.
 
@@ -94,8 +94,7 @@ No new API endpoints. `dir` is computed client-side from locale; no server chang
 - **HTML direction** — `document.documentElement.dir` set on locale change; `document.documentElement.lang` also updated.
 - **Tailwind config** — `rtl:` variant enabled in `clients/web/tailwind.config.ts`.
 - **Component audit checklist** — all layout components reviewed: Sidebar, NavBar, Breadcrumb, GradebookGrid, TipTap Editor, Modals, Dropdowns, Toasts.
-- **Storybook** — All layout stories must have an RTL story variant.
-- **Visual regression** — Chromatic or Percy snapshots for RTL + LTR to prevent regressions.
+- **Visual regression** — Playwright snapshots for RTL + LTR to prevent regressions.
 - **Text alignment** — Use `text-start` / `text-end` Tailwind classes instead of `text-left` / `text-right`.
 
 ## 11. AI / ML Considerations
@@ -109,7 +108,7 @@ None. AI-generated content may include Arabic/Hebrew text; the editor's `dir="au
 - `clients/web/src/App.tsx` — set `document.documentElement.dir` on locale change.
 - `clients/web/src/components/editor/` — TipTap RTL patches.
 - All Tailwind class usage — audit and replace physical → logical.
-- Storybook — RTL addons; visual regression integration.
+- Playwright — RTL viewport tests; visual regression integration.
 - ESLint — custom rule for physical CSS class detection.
 
 ## 13. Dependencies & Sequencing
@@ -123,7 +122,7 @@ None. AI-generated content may include Arabic/Hebrew text; the editor's `dir="au
 |---|---|---|---|
 | TipTap RTL support is incomplete upstream | H | M | File upstream issue; implement workaround with `dir="auto"` on editor node |
 | Physical property audit misses legacy CSS files | M | M | Automated grep for `margin-left/right`, `padding-left/right` in CI |
-| LTR regression after logical property conversion | M | M | Chromatic visual regression; 100% Storybook story coverage for layout components |
+| LTR regression after logical property conversion | M | M | Playwright visual regression; RTL/LTR coverage for layout components |
 
 ## 15. Rollout Plan
 
@@ -131,7 +130,7 @@ None. AI-generated content may include Arabic/Hebrew text; the editor's `dir="au
 - Phase 1: Tailwind config + `dir` switching + RTL locale registry.
 - Phase 2: Physical → logical CSS audit across all components.
 - Phase 3: TipTap RTL editor fix + directional icon mirroring.
-- Phase 4: Storybook RTL stories + visual regression baselines.
+- Phase 4: Playwright RTL snapshots + visual regression baselines.
 - GA: Arabic locale translations available + RTL layout verified.
 
 ## 16. Test Plan
@@ -147,7 +146,7 @@ None. AI-generated content may include Arabic/Hebrew text; the editor's `dir="au
 
 - Developer guide: "Building RTL-safe components with Tailwind v4 logical properties."
 - ESLint rule documentation: "Why we use `ms-*` instead of `ml-*`."
-- Storybook: RTL story template.
+- Playwright: RTL viewport test template.
 
 ## 18. Open Questions
