+++ docs/plan/21-mobile-offline-cross-platform/21.2-mobile-responsive-review.md
@@ -53,7 +53,7 @@ The Lextures web app was built desktop-first. Several critical flows are broken 
 - **FR-5.** The system MUST ensure the course navigation sidebar collapses to a bottom navigation bar or hamburger drawer on screens <768 px.
 - **FR-6.** The system MUST verify that the quiz question layout is single-column and full-width on mobile, with answer choices stacked vertically.
 - **FR-7.** The system SHOULD add `viewport` meta tag (`width=device-width, initial-scale=1`) if not already present in `index.html`.
-- **FR-8.** The system SHOULD define standard Tailwind breakpoint tokens for mobile-first design and document them in the design system (Storybook or equivalent).
+- **FR-8.** The system SHOULD define standard Tailwind breakpoint tokens for mobile-first design and document them in `clients/web/CONTRIBUTING.md`.
 
 ## 6. Non-Functional Requirements
 
@@ -140,7 +140,7 @@ Not applicable.
 
 ## 16. Test Plan
 
-- **Unit** — Storybook stories for responsive components at 375 px, 768 px viewports.
+- **Unit** — Vitest tests for responsive components at 375 px, 768 px viewports.
 - **Integration** — N/A.
 - **End-to-end** — Playwright: `isMobile: true` viewport; quiz take, gradebook scroll, module reorder.
 - **Security** — N/A.
@@ -156,7 +156,7 @@ Not applicable.
 
 ## 18. Open Questions
 
-1. Should we introduce a Storybook mobile viewport add-on for visual regression testing?
+1. Should we add Playwright mobile viewport snapshots for visual regression testing?
 2. Should the sidebar-to-bottom-nav transition be configurable by the user, or always triggered at <768 px?
 3. Is there a budget for purchasing real test devices, or do we rely on simulators/BrowserStack?
 4. Should we run Lighthouse CI on every PR or only on mobile-targeted PRs?
