+++ www/src/pages/higher-ed-page.tsx
@@ -1,5 +1,5 @@
 import { ArrowRight, BarChart3, BookOpen, BrainCircuit, GraduationCap, ShieldCheck, Unplug } from 'lucide-react'
-import { Header } from '../components/Header'
+import { Header } from '../components/header'
 
 const LINKS = {
   demo: 'https://demo.lextures.com/',
